+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPluginConstants.kt
@@ -54,7 +54,7 @@ object IntelliJPluginConstants {
         "$CACHE_REDIRECTOR/packages.jetbrains.team/maven/p/intellij-plugin-verifier/intellij-plugin-verifier"
     const val JAVA_COMPILER_ANT_TASKS_MAVEN_METADATA =
         "$DEFAULT_INTELLIJ_REPOSITORY/releases/com/jetbrains/intellij/java/java-compiler-ant-tasks/maven-metadata.xml"
-
+    const val GITHUB_REPOSITORY = "https://github.com/jetbrains/gradle-intellij-plugin"
 
     const val PLUGIN_PATH = "plugin.path"
     const val VERSION_LATEST = "latest"