+++ src/main/kotlin/org/jetbrains/intellij/tasks/DownloadRobotServerPluginTask.kt
@@ -1,6 +1,5 @@
 package org.jetbrains.intellij.tasks
 
-import org.gradle.api.GradleException
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.internal.ConventionTask
 import org.gradle.api.model.ObjectFactory
@@ -8,16 +7,13 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 import org.gradle.kotlin.dsl.property
+import org.jetbrains.intellij.IntelliJPluginConstants
 import org.jetbrains.intellij.IntelliJPluginConstants.INTELLIJ_DEPENDENCIES
-import org.jetbrains.intellij.IntelliJPluginConstants.VERSION_LATEST
 import org.jetbrains.intellij.Version
-import org.jetbrains.intellij.debug
 import org.jetbrains.intellij.logCategory
-import org.jetbrains.intellij.model.MavenMetadata
-import org.jetbrains.intellij.model.XmlExtractor
 import org.jetbrains.intellij.utils.ArchiveUtils
+import org.jetbrains.intellij.utils.LatestVersionResolver
 import java.io.File
-import java.net.URL
 import javax.inject.Inject
 
 @Suppress("UnstableApiUsage")
@@ -29,26 +25,7 @@ open class DownloadRobotServerPluginTask @Inject constructor(objectFactory: Obje
         private const val NEW_ROBOT_SERVER_DEPENDENCY = "com.intellij.remoterobot:robot-server-plugin"
         private const val NEW_ROBOT_SERVER_VERSION = "0.11.0"
 
-        fun resolveLatestVersion(): String {
-            debug(message = "Resolving latest Robot Server Plugin version")
-            val url = URL(METADATA_URL)
-            return XmlExtractor<MavenMetadata>().unmarshal(url.openStream()).versioning?.latest
-                ?: throw GradleException("Cannot resolve the latest Robot Server Plugin version")
-        }
-
-        /**
-         * Resolves the Robot Server version.
-         * If set to [org.jetbrains.intellij.IntelliJPluginConstants.VERSION_LATEST], there's request to [METADATA_URL]
-         * performed for the latest available version.
-         *
-         * @return Robot Server version
-         */
-        fun resolveVersion(version: String?) = version?.takeIf { it != VERSION_LATEST } ?: resolveLatestVersion()
-
-        fun getDependency(version: String) = when {
-            Version.parse(version) >= Version.parse(NEW_ROBOT_SERVER_VERSION) -> NEW_ROBOT_SERVER_DEPENDENCY
-            else -> OLD_ROBOT_SERVER_DEPENDENCY
-        }
+        fun resolveLatestVersion() = LatestVersionResolver.fromMaven("Robot Server Plugin", METADATA_URL)
     }
 
     @Input
@@ -68,5 +45,20 @@ open class DownloadRobotServerPluginTask @Inject constructor(objectFactory: Obje
     fun downloadPlugin() {
         val target = outputDir.get().asFile
         archiveUtils.extract(pluginArchive.get(), target, context)
+    }
+
+    /**
+     * Resolves the Robot Server version.
+     * If set to [org.jetbrains.intellij.IntelliJPluginConstants.VERSION_LATEST], there's request to [METADATA_URL]
+     * performed for the latest available version.
+     *
+     * @return Robot Server version
+     */
+    internal fun resolveRobotServerPluginVersion(version: String?) =
+        version?.takeIf { it != IntelliJPluginConstants.VERSION_LATEST } ?: resolveLatestVersion()
+
+    internal fun getDependency(version: String) = when {
+        Version.parse(version) >= Version.parse(NEW_ROBOT_SERVER_VERSION) -> NEW_ROBOT_SERVER_DEPENDENCY
+        else -> OLD_ROBOT_SERVER_DEPENDENCY
     }
 }